+++ Telegram-Mac/ModalPreviewViews.swift
@@ -314,7 +314,7 @@ class AnimatedStickerPreviewModalView : View, ModalPreviewControllerView {
             if let resource = reference.media.resource as? LocalBundleResource {
                 data = Signal { subscriber in
                     if let path = Bundle.main.path(forResource: resource.name, ofType: resource.ext), let data = try? Data(contentsOf: URL(fileURLWithPath: path), options: [.mappedRead]) {
-                        subscriber.putNext(MediaResourceData(path: path, offset: 0, size: data.count, complete: true))
+                        subscriber.putNext(MediaResourceData(path: path, offset: 0, size: Int64(data.count), complete: true))
                         subscriber.putCompletion()
                     }
                     return EmptyDisposable
